refactor: 合并个性化页面到设置页，简化项目结构
1.  将原/paint个性化页面的功能全部迁移到设置页面，移除旧的路由和相关组件
2.  更新多语言文件中导入导出的文本key
3.  升级cmzya-modern-ui到0.6.2版本以支持新的tab滚动联动功能
4.  重构设置页面布局，使用竖向tab栏分区展示所有设置项
