+++ package.json
@@ -42,7 +42,7 @@
     "@xterm/addon-fit": "^0.11.0",
     "@xterm/addon-serialize": "^0.14.0",
     "@xterm/xterm": "^6.0.0",
-    "cmzya-modern-ui": "^0.6.0",
+    "cmzya-modern-ui": "^0.6.2",
     "dompurify": "^3.3.3",
     "echarts": "^6.0.0",
     "lucide-vue-next": "^0.574.0",
